+++ adapter/taurusx/CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [1.16.3.0] - 2026-06-17
+### Changed
+- Updated SDK dependency from 1.15.2 to 1.16.3
+
 ## [1.15.2.0] - 2026-05-27
 ### Changed
 - Updated SDK dependency from 1.14.1 to 1.15.2